add boost and cuda to testUbunti and testWindows

+++ Intern/rayx-core/src/Data/xml.h
@@ -185,7 +185,7 @@ struct RAYX_API Parser {
     inline SpreadType parseEnergySpreadType() const { return static_cast<SpreadType>(parseInt("energySpreadType")); }
     inline EnergyDistributionType parseEnergyDistributionType() const { return static_cast<EnergyDistributionType>(parseInt("energyDistributionType")); }
     inline EnergySpreadUnit parseEnergySpreadUnit() const { return static_cast<EnergySpreadUnit>(parseInt("energySpreadUnit")); }
-    inline int parseNumOfEquidistantCircles() const { return parseDouble("numberCircles"); }
+    inline int parseNumOfEquidistantCircles() const { return static_cast<int>(parseDouble("numberCircles")); }
     inline Rad parseMaxOpeningAngle() const { return parseDouble("maximumOpeningAngle") / 1000.0; }
     inline Rad parseMinOpeningAngle() const { return parseDouble("minimumOpeningAngle") / 1000.0; }
     inline Rad parseDeltaOpeningAngle() const { return parseDouble("deltaOpeningAngle") / 1000.0; }
